requirements.txt
et_xmlfile==2.0.0
lxml==6.0.0
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pypandoc==1.15
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.1
rarfile==4.2
requests
six==1.17.0
tzdata==2025.1
unidecode
weasyprint==65.1
camelot-py[cv]
pdfplumber
